requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.6.0
    # via pydantic
docopt==0.6.2
    # via stomp-py
pydantic==2.7.1
pydantic-core==2.18.2
    # via pydantic
pyyaml==6.0.1
stomp-py==8.1.2
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
websocket-client==1.8.0
    # via stomp-py
